inline SubmitService implementation
